add the some function about quiz
1.add class Quiz , Question
2.the main function we use app.js
3.in questions array, we can add problem each problem structure is
(Question,choice,answer)
4.we use function “populate” to change the view for any question
5. The population function includes the procedures about change the
view.
First ->detect whether it is the end of the question. If it is ,we will
show the result how many point we get.
Second-> if it is not the result, we will getElementbyId to change the
question and answer, and set up the click event on each item